# api/openapi/l1-client.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wefty L1 client protocol",
    "version": "0.1.0-prerelease",
    "description": "Fabric-authenticated job submission and polling surface for client principals, plus the separate stable-person administrator-policy routes. Agent-only mutation verbs are in l1-agent.v1.json."
  },
  "servers": [{ "url": "wefty://control-plane" }],
  "security": [{ "fabricIdentity": [] }],
  "paths": {
...
    "/v1/jobs/{job_id}/logs": {
      "parameters": [
        {
          "name": "job_id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "getJobLogs",
        "summary": "Poll ordered job logs",
        "parameters": [
          { "name": "class", "in": "query", "required": true, "schema": { "const": "service" } },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned by the previous page.",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Log page from authoritative event rows, with an aggregate service-retention marker when history was evicted. Ordering is per attempt and stream only; next_cursor is unrelated to agent upload acknowledgements.",
            "content": {
              "application/json": {
                "schema": { "$ref": "./common.v1.json#/components/schemas/LogPage" }
              }
            }
          },
          "400": { "$ref": "./common.v1.json#/components/responses/BadRequest" },
          "401": { "$ref": "./common.v1.json#/components/responses/Unauthorized" },
          "403": { "$ref": "./common.v1.json#/components/responses/Forbidden" },
          "404": { "$ref": "./common.v1.json#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "fabricIdentity": { "$ref": "./common.v1.json#/components/securitySchemes/fabricIdentity" }
    }
  }
}